+++ changelog.md
@@ -23,6 +23,12 @@ Upgraded software (GNOME):
 - gupnp: `1.5.2 --> 1.5.3`
 - gtksourceview5: `5.5.0 --> 5.5.1`
 - rygel: `0.41.1 --> 0.41.2`
+- cups-pk-helper: `0.2.6 --> 0.2.7`
+- evolution-data-server: `3.44.3 --> 3.44.4`
+- gnome-shell: `42.3.1 --> 42.4`
+- mutter: `42.3 --> 42.4`
+- tracker: `3.3.2 --> 3.3.3`
+- zenity: `3.42.1 --> 3.43.0`
 
 # MassOS 2022.08
 First official release of MassOS with GNOME support.